move pairing functions to seperate file
