+++ src/controllers/matchmakers/pairmakers/mod.rs
@@ -1,0 +1,1 @@
+pub mod random;
